Improved the results from reports
The API expects camel cased pluralized filter params.
